agent_backup/requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-anthropic
langgraph
python-dotenv
requests
